requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.46.0
django-blacklist==0.6.1
django-crispy-forms==1.13.0
django-model-utils==4.2.0
django-notifications-hq==1.6.0
django-redis==5.1.0
django-tinymce==3.4.0
idna==3.3
jsonfield==3.1.0
mysqlclient==2.1.0
oauthlib==3.1.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
soupsieve==2.3.1
sqlparse==0.4.2
swapper==1.3.0
urllib3==1.26.7
